refactor: change bank id back to int

+++ supabase/migrations/20240405040524_create-bank-account-table.sql
@@ -1,7 +1,7 @@
 -- create a public.bank_account table for storing bank account information
 CREATE TABLE public.bank_account (
   id BIGINT NOT NULL GENERATED ALWAYS AS IDENTITY,
-  bank_id BIGINT NOT NULL REFERENCES bank ON DELETE CASCADE,
+  bank_id INTEGER NOT NULL REFERENCES bank ON DELETE CASCADE,
   org_id INTEGER NOT NULL REFERENCES org ON DELETE CASCADE,
   num TEXT,
   name TEXT,
